+++ src/main/java/pl/tw/ksiegowosc/repository/AppUserRepository.java
@@ -1,0 +1,12 @@
+package pl.tw.ksiegowosc.repository;
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import pl.tw.ksiegowosc.entity.AppUser;
+
+public interface AppUserRepository extends JpaRepository<AppUser, Long> {
+
+    Optional<AppUser> findByLogin(String login);
+}